requirements.txt
# SHL Assessment Recommendation System
# Python 3.13+

# Web Crawling
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0

# Embeddings & Vector Store
sentence-transformers>=3.3.0
faiss-cpu>=1.9.0

# LLM Integration (Groq - Free)
groq>=0.4.0

# API & Frontend
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0
pydantic>=2.10.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.67.0
numpy>=2.0.0
pymupdf>=1.24.0
